Increased buffer size from UInt16.MaxValue to Int32.MaxValue
Probably overkill but it's better than having half ban lists. ;)

+++ BattleNET/BattlEyeClient.cs
@@ -76,7 +76,7 @@ namespace BattleNET
             EndPoint remoteEP = new IPEndPoint(ipAddress, loginCredentials.Port);
 
             socket = new Socket(AddressFamily.InterNetwork, SocketType.Dgram, ProtocolType.Udp);
-            socket.ReceiveBufferSize = UInt16.MaxValue;
+            socket.ReceiveBufferSize = Int32.MaxValue;
             socket.ReceiveTimeout = 5000;
 
             try
